+++ src/schemapack/_internals/main.py
@@ -22,9 +22,9 @@ Warning: This is an internal part of the library and might change without notice
 from pathlib import Path
 
 from schemapack._internals.load import load_datapack, load_schemapack
+from schemapack._internals.validation import SchemaPackValidator
 from schemapack.spec.datapack import DataPack
 from schemapack.spec.schemapack import SchemaPack
-from schemapack.validation import SchemaPackValidator
 
 
 def load_and_validate(
