+++ src/schemapack/_internals/validation/base.py
