+++ src/schemapack/_internals/validation/plugins/content_schema.py
@@ -16,18 +16,34 @@
 
 """A validation plugin."""
 
+import json
+
+import jsonschema.exceptions
+import jsonschema.protocols
+import jsonschema.validators
+
+from schemapack._internals.validation.base import ResourceValidationPlugin
 from schemapack.exceptions import ValidationPluginError
 from schemapack.spec.custom_types import ResourceId
 from schemapack.spec.datapack import DataPack, Resource
 from schemapack.spec.schemapack import ClassDefinition
-from schemapack.validation._base import ResourceValidationPlugin
 
 
-class MultipleTargetValidationPlugin(ResourceValidationPlugin):
-    """A resource-scoped validation plugin validating the plurality of relations,
-    i.e. *_to_many (multiple.target is True) relations must be sets and *_to_one
-    (multiple.target is False) relations must be single values.
-    This only applies to schemapack classes with relations.
+def _get_json_schema_validator(schema_str: str) -> jsonschema.protocols.Validator:
+    """Get a JSON Schema validator for the given schema formatted as JSON string.
+    It is assumed that the schema has already been checked for validity against the
+    JSON Schema specs.
+    """
+    schema = json.loads(schema_str)
+    cls: type[jsonschema.protocols.Validator] = jsonschema.validators.validator_for(
+        schema
+    )
+    return cls(schema)
+
+
+class ContentSchemaValidationPlugin(ResourceValidationPlugin):
+    """A resource-scoped validation plugin validating the content of one resource
+    against the content JSON Schema defined in the corresponding schemapack.
     """
 
     @staticmethod
@@ -37,11 +53,14 @@ class MultipleTargetValidationPlugin(ResourceValidationPlugin):
 
         Returns: True if this plugin is relevant for the given class definition.
         """
-        return bool(class_.relations)
+        # Is always relevant since all resources must have a content schema:
+        return True
 
     def __init__(self, *, class_: ClassDefinition):
         """This plugin is configured with one specific class definition of a schemapack."""
-        self._relations = class_.relations
+        self._json_schema_validator = _get_json_schema_validator(
+            class_.content.json_schema
+        )
 
     def validate(
         self, *, resource: Resource, resource_id: ResourceId, datapack: DataPack
@@ -52,28 +71,9 @@ class MultipleTargetValidationPlugin(ResourceValidationPlugin):
         Raises:
             schemapack.exceptions.ValidationPluginError: If validation fails.
         """
-        wrong_relations: set[str] = set()
-        for relation_name, relation in resource.relations.items():
-            is_set = isinstance(relation, set)
-
-            try:
-                expected_set = self._relations[relation_name].multiple.target
-            except KeyError:
-                # Unknown relations are handled in a different plugin:
-                continue
-
-            if is_set != expected_set:
-                wrong_relations.add(relation_name)
-
-        if wrong_relations:
+        try:
+            self._json_schema_validator.validate(resource.content)
+        except jsonschema.exceptions.ValidationError as error:
             raise ValidationPluginError(
-                type_="CardinalityPluralityError",
-                message=(
-                    "Expected a single target ID but got a set, or vise versa, for"
-                    " the following relation propertie(s): "
-                    + ", ".join(wrong_relations)
-                ),
-                details={
-                    "wrong_relations": wrong_relations,
-                },
-            )
+                type_="ContentValidationError", message=error.message
+            ) from error
