+++ src/schemapack/_internals/validation/plugins/missing_origin.py
@@ -18,11 +18,11 @@
 
 from collections.abc import Mapping
 
+from schemapack._internals.validation.base import ClassValidationPlugin
 from schemapack.exceptions import ValidationPluginError
 from schemapack.spec.custom_types import ResourceId
 from schemapack.spec.datapack import DataPack, Resource
 from schemapack.spec.schemapack import ClassDefinition
-from schemapack.validation._base import ClassValidationPlugin
 
 
 class MissingMandatoryOriginValidationPlugin(ClassValidationPlugin):
