+++ src/schemapack/_internals/validation/plugins/one_to_many_overlap.py
@@ -19,11 +19,11 @@
 from collections import Counter
 from collections.abc import Mapping
 
+from schemapack._internals.validation.base import ClassValidationPlugin
 from schemapack.exceptions import ValidationPluginError
 from schemapack.spec.custom_types import ResourceId
 from schemapack.spec.datapack import DataPack, Resource
 from schemapack.spec.schemapack import ClassDefinition
-from schemapack.validation._base import ClassValidationPlugin
 
 
 class TargetOverlapValidationPlugin(ClassValidationPlugin):
