+++ src/schemapack/_internals/validation/plugins/unexpected_root.py
@@ -16,15 +16,16 @@
 
 """A validation plugin."""
 
+from schemapack._internals.validation.base import GlobalValidationPlugin
 from schemapack.exceptions import ValidationPluginError
 from schemapack.spec.datapack import DataPack
 from schemapack.spec.schemapack import SchemaPack
-from schemapack.validation._base import GlobalValidationPlugin
 
 
-class MissingClassSlotValidationPlugin(GlobalValidationPlugin):
-    """A global-scoped validation plugin validating that a datapack has a slot for
-    each class defined in the provided schemapack.
+class UnexpectedRootValidationPlugin(GlobalValidationPlugin):
+    """A global-scoped validation plugin validating that a datapack has no root resource
+    defined.
+    This plugin is only relevant if the schemapack has no root class defined.
     """
 
     @staticmethod
@@ -34,11 +35,11 @@ class MissingClassSlotValidationPlugin(GlobalValidationPlugin):
 
         Returns: True if this plugin is relevant for the given class definition.
         """
-        return True
+        return not bool(schemapack.root_class)
 
     def __init__(self, *, schemapack: SchemaPack):
         """This plugin is configured with the entire schemapack."""
-        self._classes = schemapack.classes.keys()
+        # there is nothing to do
 
     def validate(self, *, datapack: DataPack):
         """Validate the entire datapack.
@@ -46,15 +47,11 @@ class MissingClassSlotValidationPlugin(GlobalValidationPlugin):
         Raises:
             schemapack.exceptions.ValidationPluginError: If validation fails.
         """
-        missing_classes = [
-            class_ for class_ in self._classes if class_ not in datapack.resources
-        ]
-
-        if missing_classes:
+        if datapack.root_resource:
             raise ValidationPluginError(
-                type_="MissingClassSlotError",
-                message=("Missing slot(s) for class(es):" + ", ".join(missing_classes)),
-                details={
-                    "missing_classes": missing_classes,
-                },
+                type_="UnexpectedRootResourceError",
+                message=(
+                    "The schemapack has no root class defined but the datapack "
+                    "specifies a root resource."
+                ),
             )
