+++ src/schemapack/plugins.py
@@ -12,28 +12,18 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+#
 
-"""A specification (plus tooling) for describing linked data models based on JSON
-schema.
-"""
+"""Base classes for defining validation plugins."""
 
-from importlib.metadata import version
-
-from ._internals.isolate import isolate, isolate_class, isolate_resource
-from ._internals.load import load_datapack, load_schemapack
-from ._internals.main import load_and_validate
-from ._internals.normalize import denormalize
-from ._internals.validation import SchemaPackValidator
+from schemapack._internals.validation import (
+    ClassValidationPlugin,
+    GlobalValidationPlugin,
+    ResourceValidationPlugin,
+)
 
 __all__ = [
-    "load_and_validate",
-    "load_datapack",
-    "load_schemapack",
-    "isolate",
-    "isolate_class",
-    "isolate_resource",
-    "denormalize",
-    "SchemaPackValidator",
+    "GlobalValidationPlugin",
+    "ClassValidationPlugin",
+    "ResourceValidationPlugin",
 ]
-
-__version__ = version(__package__)
